pyproject.toml
[project]
name = "cv-perfector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=0.5.18",
    "click>=8.1.7",
    "fastapi>=0.115.4",
    "google-generativeai>=0.8.3",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.5",
    "langchain-google-genai>=2.0.4",
    "langchain>=0.3.7",
    "llama-index>=0.12.10",
    "llama-index-llms-gemini>=0.4.3",
    "loguru>=0.7.2",
    "pydantic-settings>=2.6.1",
    "pydantic[email]>=2.9.2",
    "pymupdf4llm>=0.0.17",
    "pyyaml>=6.0.2",
    "six>=1.16.0",
    "aiofiles>=24.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "deepdiff>=8.1.1",
    "rich>=13.9.4",
]
[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*Jupyter is migrating.*:DeprecationWarning",
]
